+++ lib/ts/recipe/session/types.ts
@@ -240,7 +240,7 @@ export type RecipeInterface = {
     }): Promise<SessionContainerInterface>;
     /**
      * Used to retrieve all session information for a given session handle. Can be used in place of:
-     * - getSessionData
+     * - getSessionDataFromDatabase
      * - getAccessTokenPayload
      *
      * Returns undefined if the sessionHandle does not exist
@@ -256,7 +256,11 @@ export type RecipeInterface = {
     revokeMultipleSessions(input: { sessionHandles: string[]; userContext: any }): Promise<string[]>;
 
     // Returns false if the sessionHandle does not exist
-    updateSessionData(input: { sessionHandle: string; newSessionData: any; userContext: any }): Promise<boolean>;
+    updateSessionDataInDatabase(input: {
+        sessionHandle: string;
+        newSessionData: any;
+        userContext: any;
+    }): Promise<boolean>;
 
     /**
      * @deprecated Use mergeIntoAccessTokenPayload instead
@@ -350,9 +354,9 @@ export type RecipeInterface = {
 export interface SessionContainerInterface {
     revokeSession(userContext?: any): Promise<void>;
 
-    getSessionData(userContext?: any): Promise<any>;
+    getSessionDataFromDatabase(userContext?: any): Promise<any>;
 
-    updateSessionData(newSessionData: any, userContext?: any): Promise<any>;
+    updateSessionDataInDatabase(newSessionData: any, userContext?: any): Promise<any>;
 
     getUserId(userContext?: any): string;